+++ back/cocktailpick-core/src/main/java/com/cocktailpick/core/userrecipe/dto/UserRecipeItemRequest.java
@@ -1,15 +1,16 @@
 package com.cocktailpick.core.userrecipe.dto;
 
+import javax.validation.constraints.NotBlank;
+import javax.validation.constraints.NotNull;
+
 import com.cocktailpick.core.ingredient.domain.Ingredient;
 import com.cocktailpick.core.userrecipe.domain.QuantityUnit;
 import com.cocktailpick.core.userrecipe.domain.UserRecipeItem;
+
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotNull;
 
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 @Getter
@@ -17,12 +18,12 @@ public class UserRecipeItemRequest {
     @NotNull
     private Long ingredientId;
     @NotNull
-    private int quantity;
+    private Double quantity;
     @NotBlank
     private String quantityUnit;
 
     @Builder
-    public UserRecipeItemRequest(Long ingredientId, int quantity, String quantityUnit) {
+    public UserRecipeItemRequest(Long ingredientId, Double quantity, String quantityUnit) {
         this.ingredientId = ingredientId;
         this.quantity = quantity;
         this.quantityUnit = quantityUnit;
